refactor(server): organize API routes
